CC_Camera_ov7736 the kernel meet panic when boot
BZ: 978

The var define is u16, but use read_reg(...u32), then it
will read 0(u32) to 0(u16), so other var will be cleared
with two bytes 0.

+++ drivers/media/i2c/ov7736.c
@@ -892,7 +892,7 @@ static int ov7736_s_mbus_fmt(struct v4l2_subdev *sd,
 static int ov7736_detect(struct i2c_client *client,  u16 *id, u8 *revision)
 {
 	struct i2c_adapter *adapter = client->adapter;
-	u16 high, low;
+	u32 high, low;
 	int ret;
 
 	if (!i2c_check_functionality(adapter, I2C_FUNC_I2C))
